Hide incompatible 3.5 statements
The idea is be able to run all Python versions, since 2.X to
3.5 without break the current behaviour. The Python 3.5 specific
code is hidden into private files that will be imported only in
3.5 environments.
