+++ tests/test_kernprof.py
@@ -78,28 +78,8 @@ class TestKernprof(unittest.TestCase):
         self.assertEqual(profile.enable_count, 0)
 
     if PY35:
-        def test_coroutine_decorator(self):
-            async def _():
-                async def c(x):
-                    """ A coroutine. """
-                    y = x + 10
-                    return y
-
-                profile = ContextualProfile()
-                c_wrapped = profile(c)
-                self.assertEqual(c_wrapped.__name__, c.__name__)
-                self.assertEqual(c_wrapped.__doc__, c.__doc__)
-
-                self.assertEqual(profile.enable_count, 0)
-                value = await c_wrapped(10)
-                self.assertEqual(profile.enable_count, 0)
-                self.assertEqual(value, await c(10))
-
-            import asyncio
-            loop = asyncio.get_event_loop()
-            loop.run_until_complete(_())
-            loop.close()
-
+        import _test_kernprof_py35
+        test_coroutine_decorator = _test_kernprof_py35.test_coroutine_decorator
 
 if __name__ == '__main__':
     unittest.main()
